Added `WrapperComand`, changed intro in README
`repl.base.command.WrapperCommand` uses the laziest possible adaptor to
transform python functions into a form more useful for REPL internally. The
callable is wrapper in a function that prints the return value of the callable,
and returns success whenever an exception doesn't occur.

Tweaked README blurbs in response to the existence of
[python-nubia](https://github.com/facebookincubator/python-nubia), which does
the whole CLI thing miles better than REPL does.