Actually fixed input sourcing.
Reverted to plain `input()` for `cat` and `debug`, which run as commands and
thus inside of the `sys.stdin` redirect and do not need to use their parent
REPL's `REPL.input()`. This is consistent with transparency in usage,
allowing functions with no knowledge of REPL to still work. Mostly.

This supercedes the previous couple of commits referencing routing input
through `REPL.input()`.